% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{calculate_window_limits}
\alias{calculate_window_limits}
\title{Calculates the limits of the plot window.
If the reversed_ruler is TRUE the limits are reversed. (e.g. (10, -10)
instead of (-10, 10)) so the plot is reversed. This way most of the elements
don't need to be relocated.}
\usage{
calculate_window_limits()
}
\description{
Calculates the limits of the plot window.
If the reversed_ruler is TRUE the limits are reversed. (e.g. (10, -10)
instead of (-10, 10)) so the plot is reversed. This way most of the elements
don't need to be relocated.
}
