% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_binary}
\alias{check_binary}
\title{Checks if a value is a boolean}
\usage{
check_binary(x, name)
}
\arguments{
\item{x}{The value to check}

\item{name}{Name of the parameter}
}
\description{
Checks if a value is a boolean
}
