% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_critical_diff}
\alias{check_critical_diff}
\title{Checks if the critical difference is a valid positive numeric value}
\usage{
check_critical_diff(critical_diff)
}
\arguments{
\item{critical_diff}{Numeric value}
}
\description{
Checks if the critical difference is a valid positive numeric value
}
