% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_dots_size}
\alias{check_dots_size}
\title{Checks if the dot size is a valid non-negative numeric value}
\usage{
check_dots_size(dots_size)
}
\arguments{
\item{dots_size}{Numeric value}
}
\description{
Checks if the dot size is a valid non-negative numeric value
}
