% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_input}
\alias{check_input}
\title{Checks the validity of input parameters for a statistical analysis function}
\usage{
check_input(
  means,
  labels,
  critical_diff,
  filepath,
  width,
  height,
  extra_xlim,
  extra_ylim,
  vertical,
  color_ties,
  reversed_ruler,
  show_means,
  dots_size,
  debug_
)
}
\arguments{
\item{means}{Numeric vector}

\item{labels}{Character vector}

\item{critical_diff}{Numeric value}

\item{filepath}{Character string (optional)}

\item{width}{Numeric value}

\item{height}{Numeric value}

\item{extra_xlim}{Numeric vector of length 2}

\item{extra_ylim}{Numeric vectorof length 2}

\item{vertical}{Logical}

\item{color_ties}{Logical}

\item{reversed_ruler}{Logical}

\item{show_means}{Logical}

\item{dots_size}{Numeric value}

\item{debug_}{Logical}
}
\value{
TRUE if all checks pass
}
\description{
Checks the validity of input parameters for a statistical analysis function
}
