% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_length}
\alias{check_length}
\title{Checks if a vector has the expected length}
\usage{
check_length(x, expected, name)
}
\arguments{
\item{x}{The vector to check}

\item{expected}{Expected length}

\item{name}{Name of the parameter}
}
\description{
Checks if a vector has the expected length
}
