% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_lim}
\alias{check_lim}
\title{Checks if axis limits are valid numeric vectors of length 2}
\usage{
check_lim(extra_xlim, extra_ylim)
}
\arguments{
\item{extra_xlim}{Numeric vector of length 2 for x-axis limits}

\item{extra_ylim}{Numeric vector of length 2 for y-axis limits}
}
\description{
Checks if axis limits are valid numeric vectors of length 2
}
