% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_means_and_labels}
\alias{check_means_and_labels}
\title{Validates the means and labels vectors}
\usage{
check_means_and_labels(means, labels)
}
\arguments{
\item{means}{Numeric vector}

\item{labels}{Character vector}
}
\description{
Validates the means and labels vectors
}
