% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_nonnegative_numeric}
\alias{check_nonnegative_numeric}
\title{Checks if a numeric value is non-negative and finite}
\usage{
check_nonnegative_numeric(x, name)
}
\arguments{
\item{x}{The value to check}

\item{name}{Name of the parameter}
}
\description{
Checks if a numeric value is non-negative and finite
}
