% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_numeric}
\alias{check_numeric}
\title{Checks if a value is a numeric vector without infinite values}
\usage{
check_numeric(x, name)
}
\arguments{
\item{x}{The value to check}

\item{name}{Name of the parameter}
}
\description{
Checks if a value is a numeric vector without infinite values
}
