% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_positive}
\alias{check_positive}
\title{Checks if a numeric value is strictly positive}
\usage{
check_positive(x, name)
}
\arguments{
\item{x}{The value to check}

\item{name}{Name of the parameter}
}
\description{
Checks if a numeric value is strictly positive
}
