% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{check_saving_params}
\alias{check_saving_params}
\title{Validates parameters related to file saving}
\usage{
check_saving_params(filepath, width, height)
}
\arguments{
\item{filepath}{Character string (optional)}

\item{width}{Numeric value}

\item{height}{Numeric value}
}
\description{
Validates parameters related to file saving
}
