% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{draw_critical_diff_vertical_plot}
\alias{draw_critical_diff_vertical_plot}
\title{Draws the critical difference indicator for the vertical plot, includes a
line, it limits and a text.}
\usage{
draw_critical_diff_vertical_plot()
}
\description{
Draws the critical difference indicator for the vertical plot, includes a
line, it limits and a text.
}
