% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{draw_labels_vertical_plot}
\alias{draw_labels_vertical_plot}
\title{Draw labels and the arrow which connects them to the ruler for the vertical
 plot.}
\usage{
draw_labels_vertical_plot()
}
\description{
Draw labels and the arrow which connects them to the ruler for the vertical
 plot.
}
