% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{draw_ruler_horizontal_plot}
\alias{draw_ruler_horizontal_plot}
\title{Draws the horizontal ruler, it includes the main line, integer numbers and small lines for
the ineger and the x.5 numbers.}
\usage{
draw_ruler_horizontal_plot()
}
\description{
Draws the horizontal ruler, it includes the main line, integer numbers and small lines for
the ineger and the x.5 numbers.
}
