% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{draw_tie_dots_horizontal_plot}
\alias{draw_tie_dots_horizontal_plot}
\title{Draws the points where the tie lines cross the arrows for the horizontal plot.}
\usage{
draw_tie_dots_horizontal_plot()
}
\description{
Draws the points where the tie lines cross the arrows for the horizontal plot.
}
