% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{draw_tie_lines_horizontal_plot}
\alias{draw_tie_lines_horizontal_plot}
\title{Draws the line ties between groups for the horizontal plot.}
\usage{
draw_tie_lines_horizontal_plot()
}
\description{
Draws the line ties between groups for the horizontal plot.
}
