% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{draw_tie_lines_vertical_plot}
\alias{draw_tie_lines_vertical_plot}
\title{Draws the line ties between groups for the vertical plot.}
\usage{
draw_tie_lines_vertical_plot()
}
\description{
Draws the line ties between groups for the vertical plot.
}
