% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_y_label_position_vertical_plot.R
\name{estimate_y_label_position_vertical_plot}
\alias{estimate_y_label_position_vertical_plot}
\title{Estimates the optimal y-label positions for a vertical plot}
\usage{
estimate_y_label_position_vertical_plot(p, ylim, ruler_max, ruler_min)
}
\arguments{
\item{p}{Numeric vector of initial label positions}

\item{ylim}{Numeric vector of plot y-axis limits}

\item{ruler_max}{Numeric, upper limit for label positions}

\item{ruler_min}{Numeric, lower limit for label positions}
}
\description{
Estimates the optimal y-label positions for a vertical plot
}
