% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_y_label_position_vertical_plot.R
\name{eval_f}
\alias{eval_f}
\title{Objective function for optimization, combines all penalty terms}
\usage{
eval_f(x, p, ruler_min, ruler_max, sep)
}
\arguments{
\item{x}{Numeric vector of adjusted label positions}

\item{p}{Numeric vector of initial label positions}

\item{ruler_min}{Numeric, lower limit for label positions}

\item{ruler_max}{Numeric, upper limit for label positions}

\item{sep}{Numeric, minimum required separation between labels}
}
\description{
Objective function for optimization, combines all penalty terms
}
