% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_y_label_position_vertical_plot.R
\name{eval_g_ineq}
\alias{eval_g_ineq}
\title{Penalizes ordering violations (labels appearing in the wrong order)}
\usage{
eval_g_ineq(x)
}
\arguments{
\item{x}{Numeric vector of adjusted label positions}
}
\description{
Penalizes ordering violations (labels appearing in the wrong order)
}
