% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{get_corresponding_y_coordinate_vertical_plot}
\alias{get_corresponding_y_coordinate_vertical_plot}
\title{Given a mean value, returns the corresponding y-coordinate in the vertical
plot. Without this adjustment, the coordinates would be inverted.}
\usage{
get_corresponding_y_coordinate_vertical_plot(y)
}
\arguments{
\item{y}{Numeric, the original y-coordinate}
}
\description{
Given a mean value, returns the corresponding y-coordinate in the vertical
plot. Without this adjustment, the coordinates would be inverted.
}
