% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{get_groups_data}
\alias{get_groups_data}
\title{Returns a dataframe giving the groups information of the test, where each
row is a group, the columns are:
start: the index of the means belonging to the start of the group
end: the index of the the means index belongin to the start of the group
draw_layer_col: the high layer it tie line should be draw}
\usage{
get_groups_data(means, critical_diff)
}
\arguments{
\item{means}{Vector of means to plot, the order is not necessary.}

\item{critical_diff}{Critical difference.}
}
\value{
data.frame
}
\description{
Returns a dataframe giving the groups information of the test, where each
row is a group, the columns are:
start: the index of the means belonging to the start of the group
end: the index of the the means index belongin to the start of the group
draw_layer_col: the high layer it tie line should be draw
}
