% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{get_points_coords}
\alias{get_points_coords}
\title{Returns the positions where the tie lines cross the arrows.}
\usage{
get_points_coords()
}
\description{
Returns the positions where the tie lines cross the arrows.
}
