% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{get_ties_layer}
\alias{get_ties_layer}
\title{Given a dataframe calculated in "get_groups_data" containing the groups
information of the test, where each row is a group, the columns are:
start: the index of the means belonging to the start of the group
end: the index of the the means index belongin to the start of the group}
\usage{
get_ties_layer(groups_data)
}
\arguments{
\item{groups_data}{data.frame}
}
\value{
data.frame
}
\description{
It adds the column:
draw_layer_col: the high layer it tie line should be draw
}
