% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_y_label_position_vertical_plot.R
\name{initialize_x}
\alias{initialize_x}
\title{Ensures the minimum separation distance is met on the starting point and
prevents the last element from exceeding the limit.}
\usage{
initialize_x(p, sep, ruler_min, ruler_max)
}
\arguments{
\item{p}{Numeric vector of initial label positions}

\item{sep}{Numeric, minimum required separation between labels}

\item{ruler_min}{Numeric, lower limit for label positions}

\item{ruler_max}{Numeric, upper limit for label positions}
}
\description{
Ensures the minimum separation distance is met on the starting point and
prevents the last element from exceeding the limit.
}
