% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{nemenyi_test}
\alias{nemenyi_test}
\title{Uses the funcion tsutils::nemenyi to calculate the nemenyi test}
\usage{
nemenyi_test(x, conf.level = 0.95)
}
\arguments{
\item{x}{Numeric matrix with data.}

\item{conf.level}{Confidence level.}
}
\description{
Uses the funcion tsutils::nemenyi to calculate the nemenyi test
}
