% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{nemenyi_test_and_plot}
\alias{nemenyi_test_and_plot}
\title{Calculates the Nemenyi test and plots the resuts.}
\usage{
nemenyi_test_and_plot(x, conf.level = 0.95, labels = NA, ...)
}
\arguments{
\item{x}{Numeric matrix where each row represents a measure, and each colum
a method. The labels are substracted from the column names}

\item{conf.level}{Confidence level.}

\item{labels}{Vector of labels corresponding to each means.}

\item{...}{Rest of the params from plot_nemenyi, except "means",
"critical_diff" and "labels".}
}
\description{
Calculates the Nemenyi test and plots the resuts.
}
\examples{

set.seed(1)

x = matrix( rnorm(50*6, mean = 0, sd = 1), 50, 6)

x[,1] = x[,1]+1
x[,2] = x[,2]+2
x[,3] = x[,3]+3
x[,4] = x[,4]+4
x[,5] = x[,5]+100
x[,6] = x[,5]+100

colnames(x) = c("Method A", "Method B", "Method C",
"Method D", "Method E", "Method F")

nemenyi_test_and_plot(x, conf.level=0.95)

}
