% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{nemenyi_test_and_plot_as_rows}
\alias{nemenyi_test_and_plot_as_rows}
\title{Calculates the Nemenyi test and plots the resuts. In contrast to
"nemenyi_test_and_plot" every data is stored in a row.}
\usage{
nemenyi_test_and_plot_as_rows(d, score, method, group, conf.level = 0.95, ...)
}
\arguments{
\item{d}{Data frame containing the data.}

\item{score}{Character name of the column representing the score values.}

\item{method}{Character name of the column representing the method.}

\item{group}{Character name of the column representing the grouping variable}

\item{conf.level}{Confidence level.}

\item{...}{Rest of the params from plot_nemenyi, except "means",
"critical_diff" and "labels".}
}
\description{
Calculates the Nemenyi test and plots the resuts. In contrast to
"nemenyi_test_and_plot" every data is stored in a row.
}
\examples{

set.seed(1)

methods = c("Method A","Method B","Method C",
"Method D", "Method E", "Method F")

df = expand.grid(
  fold = 1:10,
  method = methods
)

df$acc = c(
  rnorm(10,mean=0,sd=1),
  rnorm(10,mean=1,sd=1),
  rnorm(10,mean=2,sd=1),
  rnorm(10,mean=3,sd=1),
  rnorm(10,mean=100,sd=1),
  rnorm(10,mean=200,sd=1)
)

nemenyi_test_and_plot_as_rows(df, "acc", "method", "fold")

}
