% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{numbers_2_str_same_rounding}
\alias{numbers_2_str_same_rounding}
\title{Casts a vector of number to string keeping the maximum number of decimals for
all the numbers.}
\usage{
numbers_2_str_same_rounding(numbers)
}
\arguments{
\item{numbers}{Numeric vector}
}
\description{
Casts a vector of number to string keeping the maximum number of decimals for
all the numbers.
}
