% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_y_label_position_vertical_plot.R
\name{penalization_change}
\alias{penalization_change}
\title{Penalizes changes from the original positions}
\usage{
penalization_change(x, p)
}
\arguments{
\item{x}{Numeric vector of adjusted label positions}

\item{p}{Numeric vector of initial label positions}
}
\description{
Penalizes changes from the original positions
}
