% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_y_label_position_vertical_plot.R
\name{penalization_separation}
\alias{penalization_separation}
\title{Penalizes label positions that do not maintain the minimum separation}
\usage{
penalization_separation(x, sep)
}
\arguments{
\item{x}{Numeric vector of adjusted label positions}

\item{sep}{Numeric, minimum required separation between labels}
}
\description{
Penalizes label positions that do not maintain the minimum separation
}
