% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{plot_nemenyi}
\alias{plot_nemenyi}
\title{Generates a plot for a Nemenyi test results.}
\usage{
plot_nemenyi(
  means,
  labels,
  critical_diff,
  filepath = NA,
  width = NA,
  height = NA,
  extra_xlim = c(0, 0),
  extra_ylim = c(0, 0),
  vertical = F,
  color_ties = T,
  reversed_ruler = F,
  show_means = F,
  dots_size = 1,
  debug_ = F
)
}
\arguments{
\item{means}{Vector of means to plot, the order is not necessary.}

\item{labels}{Vector of labels corresponding to each means.}

\item{critical_diff}{Critical difference.}

\item{filepath}{(Optional) File path to save the plot.}

\item{width}{(Optional) If filepath is provided, width of the plot to save.}

\item{height}{(Optional) If filepath is provided, height of the plot to save.}

\item{extra_xlim}{(Optional) Left and right extra space in the x-axis.}

\item{extra_ylim}{(Optional) Bottom and top extra space in the y-axis.}

\item{vertical}{(Optional) If TRUE the rule is vertical}

\item{color_ties}{(Optional) If TRUE the tie lines are colored.}

\item{reversed_ruler}{(Optional) If TRUE the ruler is reversed.}

\item{show_means}{(Optional) If TRUE shows the means in the labels.}

\item{dots_size}{(Optional) Dots size, if zero dots are not shown.}

\item{debug_}{(Optional) If TRUE plots debug information.}
}
\description{
Generates a plot for a Nemenyi test results.
}
\examples{
library(NemenyiPlot)

cd = 1

means = c(1, 1.01, 1.2, 3, 3.2, 3.5, 3.5, 3.5, 3.5, 6, 7)

labels = c(
  "Method E",
  "Method K",
  "Method D",
  "Method B",
  "Method C",
  "Method J",
  "Method A",
  "Method G",
  "Method I",
  "Method F",
  "Method H"
)

NemenyiPlot::plot_nemenyi(means, labels, cd)
NemenyiPlot::plot_nemenyi(means, labels, cd,
  reversed_ruler = TRUE, dots_size = 0)
NemenyiPlot::plot_nemenyi(means, labels, cd,
  vertical = FALSE, show_means = TRUE)
NemenyiPlot::plot_nemenyi(means, labels, cd,
 vertical = FALSE, reversed_ruler = TRUE, color_ties = FALSE)

}
