% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi_plot.R
\name{remove_non_first_group_item}
\alias{remove_non_first_group_item}
\title{Removes the non first item in groups, from a data.frame where each row is a
group and the columns are:
start: the index of the means belonging to the start of the group
end: the index of the the means index belongin to the start of the group}
\usage{
remove_non_first_group_item(groups_data)
}
\arguments{
\item{groups_data}{data.frame}
}
\description{
Removes the non first item in groups, from a data.frame where each row is a
group and the columns are:
start: the index of the means belonging to the start of the group
end: the index of the the means index belongin to the start of the group
}
